// src/model/art.ts
export type Art = {
    art_id: number;
    type_id: number;
    type_id_1: number;
    type_name: string;
    group_id: number;
    art_name: string;
    art_sub: string;
    art_author: string;
    art_tag: string;
    art_pic: string;
    art_blurb: string;
    art_time: string;
    art_content: string;
};